refactor(v0.14.13): shared AttributeSelect + RatingNeighbors — deselect, bulk/media unify, neighbor helper in bulk
Extracts the rating-entry UI duplicated inline across RatingCard and
BulkRateDialog into two shared components so they can't drift:

- AttributeSelect: the single attribute dropdown for both. Adds a ✕ clear
  button to deselect an attribute back to "Not set", and filters the
  `not_applicable` sentinel out of the selectable list everywhere — it's now
  auto-set-only (never user-pickable), which closes the bulk-vs-media drift
  (bulk used to offer "Not Applicable"). The trigger keeps no extra right
  padding so the built-in chevron stays at the edge; the ✕ is an absolute
  overlay just left of it.
- RatingNeighbors: the "How you rated similar titles" panel, now mounted in
  the bulk dialog too (anime-scoped, so it gets the anime's
  uuid/genres/studios/age as exclude+tiebreak context, like the media page).
  The bulk dialog body gets `min-w-0` so long neighbor titles truncate instead
  of overflowing (Dialog.Content is a CSS grid), and the dialog is widened.

Also adds `attributeBadges` (utils/ratingAttributes.ts, shared by the filled
display + neighbor rows) and `formatScore` (utils/formatString.ts), which
rounds a rating to 2dp before formatting so float noise like
7.890000000000001 renders as 7.89.
