+++ phsar/frontend/USER_FLOWS.md
@@ -199,7 +199,7 @@ Each anime search result card shows:
   - Clicking rows toggles selection instead of navigating
   - Action bar slides in: "Select all/Deselect all", selected count, "Rate" button, "Watchlist" button
   - "Delete Ratings" button appears when any selected media have existing ratings
-  - **Rate** opens BulkRateDialog: score circle + slider, note (applied to last main media), collapsible attributes grid. Overwrite warning shown if any selected media already rated. On save: exits select mode, shows "Note Added" info dialog naming which media received the note.
+  - **Rate** opens BulkRateDialog: score circle + slider, note (applied to last main media), collapsible attributes grid (each attribute clearable via a ✕), and the same "How you rated similar titles" panel as the media page (bulk is anime-scoped, so neighbors come from other anime). Overwrite warning shown if any selected media already rated. On save: exits select mode, shows "Note Added" info dialog naming which media received the note.
   - **Delete Ratings** opens a destructive confirmation dialog. When any selected media have recorded watches, the dialog offers an "Also delete watch history" checkbox showing how many have watches (and how many were watched more than once); kept by default. Submits `POST /ratings/bulk-delete?delete_watch_history=`
   - **Watchlist** opens a stub dialog (wired in v0.15.0)
   - "Cancel" exits select mode and clears selection
@@ -258,8 +258,8 @@ Each anime search result card shows:
   - Score: editable circle with direct text input + slider (0-10, step 0.5)
   - Watch-status selector (segmented: Completed / On Hold / Dropped) + episodes watched input (auto-filled with total episodes when Completed; revealed/editable when On Hold or Dropped)
   - Note textarea (max 1000 chars with counter)
-  - Collapsible "Details" section with 11 attribute selectors (pace, animation quality, 3D animation, watched format, fan service, dialogue quality, character depth, ending type, ending quality, story quality, originality) — shows set/total count badge. Ending quality: when On Hold or Dropped, auto-set to "Not Applicable" and disabled; when Completed, only 3 quality options shown (Unsatisfying, Satisfying, Exceptional)
-  - Collapsible "How you rated similar titles" panel — on first expand fetches the user's ratings once (`GET /ratings/scores`) and shows the 2 closest at-or-below + 2 closest above the current score, each from a different anime (only **completed** ratings, so dropped/on-hold scores don't skew the comparison); the list re-selects live as the score slider moves (no refetch). Titles render in the user's name language; each row shows cover + title + score and expands to reveal that title's attributes. Empty/short state prompts the user to rate more titles
+  - Collapsible "Details" section with 11 attribute selectors (pace, animation quality, 3D animation, watched format, fan service, dialogue quality, character depth, ending type, ending quality, story quality, originality) — shows set/total count badge. Each selector has a ✕ to clear it back to "Not set". Ending quality: when On Hold or Dropped it's auto-set to "Not Applicable" and disabled; otherwise the options are Unsatisfying / Satisfying / Very Satisfying — "Not Applicable" is never user-selectable (an auto-set-only sentinel)
+  - Collapsible "How you rated similar titles" panel (shared by the media rating card and the bulk dialog) — on first expand fetches the user's ratings once (`GET /ratings/scores`) and shows the 2 closest at-or-below + 2 closest above the current score, each from a different anime (only **completed** ratings, so dropped/on-hold scores don't skew the comparison); the list re-selects live as the score slider moves (no refetch). Titles render in the user's name language; each row shows cover + title + score and expands to reveal that title's attributes. Empty/short state prompts the user to rate more titles
   - Submit/Update button (disabled when no changes detected on existing rating)
   - **Downgrade prompt**: saving a change from Completed to On Hold/Dropped while watch history exists opens a "Keep your watch history?" pop-up — Keep history / Remove history / Cancel.
   - Cancel button returns to display mode
